+++ packages/qwik/src/adapters/zodField$.ts
@@ -1,15 +1,16 @@
 import { $, implicit$FirstArg, type QRL } from '@builder.io/qwik';
 import type { ZodType } from 'zod';
-import type { FieldValue, Maybe, ValidateField } from '../types';
+import { getParsedZodSchema } from '../utils';
+import type { FieldValue, Maybe, MaybeFunction, ValidateField } from '../types';
 
 /**
  * See {@link zodField$}
  */
 export function zodFieldQrl<TFieldValue extends FieldValue>(
-  schema: QRL<ZodType<any, any, TFieldValue>>
+  schema: QRL<MaybeFunction<ZodType<any, any, TFieldValue>>>
 ): QRL<ValidateField<TFieldValue>> {
   return $(async (value: Maybe<TFieldValue>) => {
-    const result = (await schema.resolve()).safeParse(value);
+    const result = await getParsedZodSchema(schema, value);
     return result.success ? '' : result.error.issues[0].message;
   });
 }