+++ packages/qwik/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to the library will be documented in this file.
 
+## vX.X.X (Month DD, YYYY)
+
+- Extend parameter of `zodField$` and `zodForm$` to allow passing a function (issue #38)
+
 ## v0.3.1 (March 24, 2023)
 
 - Refactor processing of numbers in `getElementInput` util (issue #34)